Move cardano-wallet launching code into its own module
